Remove practically unreachable code
* diff-varint-snappy postings are now only used for Series calls
  * this commit moves diffVarintSnappyDecode to testing code
  * removes being able to handle diff-varint-snappy postings without matchers
* since now caching log expects to have some anti-collision material in the cache item,
  we can also remove the handling of raw stored postings. I checked the metric for compression
  failures (cortex_bucket_store_cached_postings_compression_errors_total)
  in the last 30 days across all grafana labs clusters, and it is always 0
